% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_calc.R
\name{calc.lpi}
\alias{calc.lpi}
\title{The main processing function for the module}
\usage{
calc.lpi(
  dbname,
  host,
  port,
  user,
  password,
  indicator.path = NULL,
  test = NULL,
  out.file = NULL,
  calc.lpisep = ",",
  enable_parallel = FALSE
)
}
\arguments{
\item{dbname}{A string. The database name to connect to in the postgres
instance.}

\item{host}{A string. The IP address or DNS name which hosts the database.}

\item{port}{An integer. The port which the postgres service monitors for
connections.}

\item{user}{A string. The database user used to connect to the database.}

\item{password}{A string. The password used to connect to the database.}

\item{indicator.path}{A string. The file path to a tab delimited file which
contains indicator definitions. See README for file requirements.}

\item{test}{A string. The file path to a directory in which to output a
filtered version of the raw data for each indicator, for testing purposes.}

\item{out.file}{A string. The file path for the output file. Must have a .rds
or .csv extension.}

\item{enable_parallel}{A Boolean flag telling the system to run parallel
processing for those calculations and system on which it is supported. This
can be useful generally for smaller data sets with many indicators.}

\item{sep}{A character which is used to delimit the output file
(\code{out.file}) in the case that it is to be in the CSV format.}
}
\value{
A list containing process time elapsed information.
}
\description{
The main processing function for the module
}
