% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_calc.R
\name{calc.plot}
\alias{calc.plot}
\title{This function takes calculated indicators at the survey level (reckey) and
adds zeros (cover) or NA's (height) for those surveys which do not have
values for specific indicators and then calculates the average for each
unique plot/survey year.}
\usage{
calc.plot(imported, indicators, indicator.list = NULL)
}
\arguments{
\item{imported}{A list of tibbles produced from \code{import.data()}}

\item{indicators}{A tibble produced from \code{calc.indicators()}. These are
the calculated indicators at the survey (reckey) level.}

\item{indicator.list}{A tibble produced from \code{load.indicators()}}
}
\value{
A tibble containing the calculated indicators averaged at the plot
level.
}
\description{
This function takes calculated indicators at the survey level (reckey) and
adds zeros (cover) or NA's (height) for those surveys which do not have
values for specific indicators and then calculates the average for each
unique plot/survey year.
}
