% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{check.compatible}
\alias{check.compatible}
\title{Checks if a data source can be read by the import functions.}
\usage{
check.compatible(path)
}
\arguments{
\item{path}{character string. The file path to the data source.}
}
\value{
A list with compatibility characteristics.
}
\description{
Checks if a data source can be read by the import functions.
}
