% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{comment.from.xml}
\alias{comment.from.xml}
\title{Main processing function for inserting iso19110-2016 catalog metadata in XML
format into the database as comments.}
\usage{
comment.from.xml(xml.path)
}
\arguments{
\item{xml.path}{A character string. Path to the xml metadata for the
database.}
}
\description{
Main processing function for inserting iso19110-2016 catalog metadata in XML
format into the database as comments.
}
