% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{compare.types}
\alias{compare.types}
\title{Determines which columns have different types between a data source and
destination. and
type}
\usage{
compare.types(src.cols, dest.cols)
}
\arguments{
\item{src.cols}{A tibble with two fields, column_name (string) and
data_type (string), the former containing column names from source tibble
tables and the latter denoting their R data type (e.g. integer, logical,
etc.)}

\item{dest.cols}{A tibble with two fields, column_name (string) and
data_type (string), the former containing column names from Postgres
destination table and the latter denoting their Postgres data type (e.g.
character varying, double precision, etc.)}
}
\value{
A logical vector with TRUE indicating same column type and FALSE
indicating otherwise.
}
\description{
Determines which columns have different types between a data source and
destination. and
type
}
