% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{convert.sqlite}
\alias{convert.sqlite}
\title{Takes a CREATE TABLE SQL statement for a Postgres instance and converts it
to A CREATE TABLE statement that will work for SQLite.}
\usage{
convert.sqlite(sql)
}
\arguments{
\item{sql}{A string CREATE TABLE statement for a Postgres instance.}
}
\value{
A list containing a string CREATE TABLE statement for an SQLite
database, the name of the table, and geometry type and srid variables if
present.
}
\description{
Takes a CREATE TABLE SQL statement for a Postgres instance and converts it
to A CREATE TABLE statement that will work for SQLite.
}
