% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{copy.post.tables}
\alias{copy.post.tables}
\title{Queries the Postgres instance for table data and inserts data into an SQLite
table.}
\usage{
copy.post.tables(schema, spatial)
}
\arguments{
\item{schema}{A string. The name of the schema in the Postgres instance.}

\item{spatial}{logical. A flag tell the function that \code{table} contains
PostGIS geometry columns.}
}
\description{
Queries the Postgres instance for table data and inserts data into an SQLite
table.
}
