% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create.comment.sql}
\alias{create.comment.sql}
\title{Creates comment SQL.}
\usage{
create.comment.sql(tbl.type, schema, name, def, col = NULL)
}
\arguments{
\item{tbl.type}{character string. The type of comment to be created. One of
c("table", "view", "column")}

\item{schema}{character string. The name of the schema on which to comment.}

\item{name}{character string. The name of the table or view on which to
comment.}

\item{def}{character string. The comment text.}

\item{col}{character string. The name of the column on which to comment.
Can be NULL to make a table or view comment.}
}
\value{
character string. An SQL statement that can be executed by the
database.
}
\description{
Creates comment SQL.
}
