% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create.dima}
\alias{create.dima}
\title{Creates, loads data into, and executes other statements within dima schema
of the Postgres instance.}
\usage{
create.dima(sql.path)
}
\arguments{
\item{sql.path}{A string directory path which points to the location
where SQL statements to be executed on the PostGIS instance are stored.}
}
\description{
Creates, loads data into, and executes other statements within dima schema
of the Postgres instance.
}
