% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create.exts}
\alias{create.exts}
\title{Creates extensions on the public schema of the Postgres instance.}
\usage{
create.exts(sql.path)
}
\arguments{
\item{sql.path}{A string directory path which points to the location
where SQL statements to be executed on the PostGIS instance are stored.}
}
\description{
Creates extensions on the public schema of the Postgres instance.
}
