% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_convert.R
\name{create.filter.df}
\alias{create.filter.df}
\title{Creates a filter.df tibble to be used in \code{lpi.to.gis} from strings.
Primarily used for command line purposes.}
\usage{
create.filter.df(cover_str = NULL, height_str = NULL, dead_str = NULL)
}
\arguments{
\item{cover_str}{character string. A semicolon delimited string of comma
delimited pairs consisting of \code{indicator} and \code{hit_type} that
should be selected from a \code{calc.lpi} output which will be used to
create a long -> wide sf data frame for export using cover fields
(cvr_*).}

\item{height_str}{character string. A semicolon delimited string of comma
delimited pairs consisting of \code{indicator} and \code{hit_type} that
should be selected from a \code{calc.lpi} output which will be used to
create a long -> wide sf data frame for export using height fields
(height_*).}

\item{dead_str}{character string. A semicolon delimited string of comma
delimited pairs consisting of \code{indicator} and \code{hit_type} that
should be selected from a \code{calc.lpi} output which will be used to
create a long -> wide sf data frame for export using dead fields
(dead_*).}
}
\value{
a tibble
}
\description{
Creates a filter.df tibble to be used in \code{lpi.to.gis} from strings.
Primarily used for command line purposes.
}
\examples{
create.filter.df(cover_str = "indicator1, any; indicator2, top")
}
