% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{create.insert}
\alias{create.insert}
\title{Will create an INSERT SQL statement for a Postgres table based on the
contents of the data source.}
\usage{
create.insert(
  schema,
  table.name,
  named = FALSE,
  update = FALSE,
  ins.cols = NULL,
  update.cols = NULL,
  cast = NULL,
  srid = 4326
)
}
\arguments{
\item{schema}{A string. The name of the schema in the destination.}

\item{table.name}{A name of the table in the destination for which to build the INSERT
statement}

\item{named}{logical. A flag that tells the function to build the SQL based
on glue::glue syntax (e.g. {my.parameter}) or parameter position syntax
(e.g. $1, $2, etc.)}

\item{update}{logical. A flag that instructs the function to construct an
UPSERT statement instead of an INSERT statement.}

\item{ins.cols}{A string vector containing the names of columns to use for
the insert statement.}

\item{update.cols}{A string vector containing the names of columns to UPDATE
in the case of an upsert statement.}

\item{cast}{A logical vector telling the function whether to encapsulate the
parameter in the SQL in a CAST function, hopefully mitigating type errors.}

\item{srid}{An integer denoting the SRID (EPSG code) to use when inserting
geometry data into tables that support PostGIS geometry.}
}
\value{
a list containing the built INSERT SQL and the list of columns used
to build it.
}
\description{
Will create an INSERT SQL statement for a Postgres table based on the
contents of the data source.
}
