% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create.lleiadb}
\alias{create.lleiadb}
\title{Creates a PostGIS database and populates it with default data.}
\usage{
create.lleiadb(
  dbname,
  user,
  password = getPass::getPass(),
  sql.path = "sql",
  spatial.path = "spatial",
  host = "localhost",
  port = 5432
)
}
\arguments{
\item{dbname}{character vector. The database name to connect to in the
postgres instance.}

\item{user}{character vector. The database user used to connect to the
database.}

\item{password}{(optional) character vector. The password used to connect to
the database. If no password is provided, user will be prompted upon
execution.}

\item{sql.path}{(optional) A character vector directory path which points to
the location where the SQL files for DB creation are stored
(default = "sql").}

\item{spatial.path}{(optional) A character vector directory path which
points to the location where spatial data to be imported into the PostGIS
instance is stored (default = "spatial").}

\item{host}{(optional) character vector. The IP address or DNS name which
hosts the database (default = "localhost").}

\item{port}{(optional) An integer. The port which the postgres service
monitors for connections (default = 5432).}
}
\description{
Creates a PostGIS database and populates it with default data.
}
