% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create.lmf}
\alias{create.lmf}
\title{Creates, loads data into, and executes other statements within lmf schema
of the Postgres instance.}
\usage{
create.lmf(sql.path)
}
\arguments{
\item{sql.path}{A character vector directory path which points to the
location where SQL statements to be executed on the PostGIS instance are
stored.}
}
\description{
Creates, loads data into, and executes other statements within lmf schema
of the Postgres instance.
}
