% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create.post.select}
\alias{create.post.select}
\title{Creates an SQLite SELECT statement based on the table definition present
in the Postgres instance.}
\usage{
create.post.select(schema, table, spatial)
}
\arguments{
\item{schema}{A string. The name of the schema in the Postgres instance.}

\item{table}{A string. The name of the table in the Postgres instance.}

\item{spatial}{logical. A flag tell the function that \code{table} contains
PostGIS geometry columns.}
}
\value{
A list containing the SELECT SQL statement and a vector of the
column names used to construct it.
}
\description{
Creates an SQLite SELECT statement based on the table definition present
in the Postgres instance.
}
