% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create.spatial}
\alias{create.spatial}
\title{Imports spatial features into the PostGIS database.}
\usage{
create.spatial(spatial.path)
}
\arguments{
\item{spatial.path}{A character vector directory path which points to the
location where spatial data to be imported into the PostGIS instance is
stored.}
}
\description{
Imports spatial features into the PostGIS database.
}
