% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create.sqlite.insert}
\alias{create.sqlite.insert}
\title{This function creates an SQLite INSERT statement based on the table
definition stored in the Postgres instance.}
\usage{
create.sqlite.insert(
  schema,
  table,
  update = FALSE,
  ins.cols = NULL,
  srid = 4326
)
}
\arguments{
\item{schema}{A string. The name of the schema in the Postgres instance.}

\item{table}{A string. The name of the table in the Postgres instance.}

\item{update}{logical. A flag telling the function to create an UPSERT
statement rather than an INSERT statement.}

\item{ins.cols}{A string vector which contains columns names to use in the
INSERT statement.}

\item{srid}{An integer which is the srid/EPSG code representing the GPS
coordinate system/datum to store spatial data in.}
}
\value{
A list containing the INSERT SQL statement as well as a vector of
column names used to construct it.
}
\description{
This function creates an SQLite INSERT statement based on the table
definition stored in the Postgres instance.
}
