% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{create.sqlite}
\alias{create.sqlite}
\title{Loads CREATE TABLE statements from a file, separates them, converts them to
SQlite statements, executes them and adds spatial metadata and columns if
necessary}
\usage{
create.sqlite(con, spatial, sql.path)
}
\arguments{
\item{con}{An RSQLite database connection.}

\item{spatial}{logical. A flag telling the function to initialize SpatiaLite
metadata, converting the database from SQLite to SpatialLite.}

\item{sql.path}{A string file path to a file containing Postgres CREATE TABLE
statements.}
}
\description{
Loads CREATE TABLE statements from a file, separates them, converts them to
SQlite statements, executes them and adds spatial metadata and columns if
necessary
}
