% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{create.sqlite.select}
\alias{create.sqlite.select}
\title{Constructs a SELECT statment which can be used to retrieve source data from
A spatially enabled sqlite database by converting table geometry to WKT
format and numeric DATETIME columns to TIMESTAMP compatible format.
The function assumes that datetime values are stores numerically in the
sqlite database because that is the field type sqlite assumes when a table
CREATE statement uses a data type of TIMESTAMP, DATE, OR DATETIME.}
\usage{
create.sqlite.select(sqlite.con, table, spatial)
}
\arguments{
\item{sqlite.con}{An \code{Rsqlite} connection object.}

\item{table}{A string. The name of the table to build the select statement
for.}

\item{spatial}{logical. A flag which tells the function to convert geometry
columns to WKT format for import.  The geometry must be named "geom" in
the source data.}
}
\value{
A list consisting of the constructed SELECT statement and the fields
used to construct it.
}
\description{
Constructs a SELECT statment which can be used to retrieve source data from
A spatially enabled sqlite database by converting table geometry to WKT
format and numeric DATETIME columns to TIMESTAMP compatible format.
The function assumes that datetime values are stores numerically in the
sqlite database because that is the field type sqlite assumes when a table
CREATE statement uses a data type of TIMESTAMP, DATE, OR DATETIME.
}
