% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dbbind.insert.rw}
\alias{dbbind.insert.rw}
\title{This is an alternative row-wise insert function using \code{DBI::dbBind()}
instead of glue::glue functionality like \code{rowwise.insert()}.
CURRENTLY NOT IMPLEMENTED but kept for reference.}
\usage{
dbbind.insert.rw(table, insert)
}
\arguments{
\item{table}{A tibble containing the source data to be inserted.}

\item{insert}{A list constructed via \code{create.insert()}.}
}
\value{
The number of rows affected by the insert.
}
\description{
This is an alternative row-wise insert function using \code{DBI::dbBind()}
instead of glue::glue functionality like \code{rowwise.insert()}.
CURRENTLY NOT IMPLEMENTED but kept for reference.
}
