% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_calc.R
\name{do.indicator.calc}
\alias{do.indicator.calc}
\title{This function takes an indicator list, and runs \code{calc.indicators()} on
every indicator in the list, then binds the rows together into a final table.}
\usage{
do.indicator.calc(
  imported,
  indicator.list = NULL,
  num.cores = 1,
  use.mc = FALSE
)
}
\arguments{
\item{imported}{A list of tibbles produced from \code{import.data()}}

\item{indicator.list}{A tibble produced from \code{load.indicators()}}

\item{num.cores}{The number of cpu cores to use to do the indicator
calculation. Optimal number is \code{nrow(indicator.list}).}

\item{use.mc}{logical. A flag to tell the function whether or not to use
multicore processing.}
}
\value{
A tibble containing rows for every indicator processed for each
unique plotkey and survey year.
}
\description{
This function takes an indicator list, and runs \code{calc.indicators()} on
every indicator in the list, then binds the rows together into a final table.
}
