% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_calc.R
\name{do.indicator.test}
\alias{do.indicator.test}
\title{Takes input from a user provided indicator list and exports the filtered data
(either in long or wide form) into individual delimited files in order to
test that the indicator filter string is working as expected.}
\usage{
do.indicator.test(imported, indicator.list, test.dir, sep = ",")
}
\arguments{
\item{imported}{A list of tibbles produced from \code{import.data()}}

\item{indicator.list}{A tibble produced from \code{load.indicators()}}

\item{test.dir}{A string directory file path in which to save
the individual indicator output files.}

\item{sep}{A character. The character to use whe writing the delimited
outputs.}
}
\description{
Takes input from a user provided indicator list and exports the filtered data
(either in long or wide form) into individual delimited files in order to
test that the indicator filter string is working as expected.
}
