% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{execute.comment}
\alias{execute.comment}
\title{An SQL execution function which uses a try-catch statement to skip/print
errors if found in the comment SQL.}
\usage{
execute.comment(sql)
}
\arguments{
\item{sql}{character string. The SQL to execute.}
}
\description{
An SQL execution function which uses a try-catch statement to skip/print
errors if found in the comment SQL.
}
