% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_convert.R
\name{filter.table}
\alias{filter.table}
\title{takes a filter string and applies the filter to the raw input data, in the
case of multiple filters, uses dplyr::bind_rows() to combine.}
\usage{
filter.table(in.tbl, val, filter.str)
}
\arguments{
\item{in.tbl}{A tibble. Produced by the lpi_calc module and imported.}

\item{val}{A string. Determines which field is pivoted longer.
One of c("cover", "height", "dead")}

\item{filter.str}{A string in the format of "{indicator1}, {hit_type1};
{indicator2}, {hit_type2}; etc." which is used to build filters for in.tbl}
}
\value{
A tibble which has been pivoted to a wide format via
tidyr::pivot_longer() where each row has a unique plotkey and survey_year
}
\description{
takes a filter string and applies the filter to the raw input data, in the
case of multiple filters, uses dplyr::bind_rows() to combine.
}
