% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_convert.R
\name{filter.widen}
\alias{filter.widen}
\title{Takes a filter string and applies the filter to the raw input data, in the
case of multiple filters, uses dplyr::bind_rows() to combine.}
\usage{
filter.widen(in.tbl, filter.df, sd = TRUE, n = TRUE)
}
\arguments{
\item{in.tbl}{A tibble. Produced by the lpi_calc module and imported.}

\item{filter.df}{A tibble with the character fields: \code{field},
\code{indicator}, and \code{hit_type}. The \code{field} column must contain
one of \code{c("cover", "height", "dead")}. The indicator field must match
indicators to filter \code{in.tbl} by or NA, which will result in all
indicators of a certain \code{hit_type} being selected. The same logic
applies to the \code{hit_type} field.}

\item{sd}{Boolean. Use standard deviation field(s) in wide result.}

\item{n}{Boolean. Use line count (n) field(s) in wide result.}
}
\value{
A tibble which has been pivoted to a wide format via
tidyr::pivot_longer() where each row has a unique plotkey and survey_year
}
\description{
Takes a filter string and applies the filter to the raw input data, in the
case of multiple filters, uses dplyr::bind_rows() to combine.
}
