% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get.dest.info}
\alias{get.dest.info}
\title{Gets column names from a Postgres schema/table.}
\usage{
get.dest.info(schema, table)
}
\arguments{
\item{schema}{A string. The schema that contains the table to return info
about.}

\item{table}{A string. The table name to return info about.}
}
\value{
A list of tibbles, one containing column information, and another
containing constraint information.
}
\description{
Gets column names from a Postgres schema/table.
}
