% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get.dest.tables}
\alias{get.dest.tables}
\title{Queries the Postgres LLEIA db instance for tables in a specific schema}
\usage{
get.dest.tables(schema)
}
\arguments{
\item{schema}{A string. The schema to query for tables.}
}
\value{
A table containing the table names in \code{schema}.
}
\description{
Queries the Postgres LLEIA db instance for tables in a specific schema
}
