% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get.info}
\alias{get.info}
\title{Gathers information about both source and destination tables and determines
which fields are matched in name and type.}
\usage{
get.info(schema, tbl.name, src.data)
}
\arguments{
\item{schema}{A string. The name of the schema to be queried in the Postgres
connection.}

\item{tbl.name}{A string. The name of the table to retrieve information about
from the postgres connection.}

\item{src.data}{A tibble containing the source data having the name of
\code{tbl.name}}
}
\value{
A list of vectors containing information about missing columns in the
destination that are in the source, missing columns in the source that are
in the destination, columns which match in both source and destination,
and whether or not those columns need to be CAST during insert.
}
\description{
Gathers information about both source and destination tables and determines
which fields are matched in name and type.
}
