% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get.key}
\alias{get.key}
\title{This function generates an md5hash for database files or ESRI file
geodatabase folders.}
\usage{
get.key(src)
}
\arguments{
\item{src}{A string file path to a file with one of the following extensions:
(.mdb, .accdb, .sqlite, .db, .gpkg) or a folder path with the .gdb
extension.}
}
\value{
An md5hash of the file, or in the case of a folder, a hash of the
vector of hashes for each of the files in the folder, minus certain files
in the gdb folder which are mutable or transient.
}
\description{
This function generates an md5hash for database files or ESRI file
geodatabase folders.
}
