% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get.src.info}
\alias{get.src.info}
\title{Returns information about field types for imported source data.}
\usage{
get.src.info(tbl)
}
\arguments{
\item{tbl}{A tibble containing the source data.}
}
\value{
A tibble containing column information for \code{tbl}.
}
\description{
Returns information about field types for imported source data.
}
