% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get.src.tables}
\alias{get.src.tables}
\title{Retrieves data tables from data source and returns them as a list of tibbles.}
\usage{
get.src.tables(path, md5hash, key, desc = NULL)
}
\arguments{
\item{path}{A string file path pointing to the source database. File name
extensions must be one of (.mdb, .accdb, .sqlite, .db) and folder paths
must have the extension of (.gdb).}

\item{md5hash}{A string. The md5hash of the database, produced via
\code{get.key()}}

\item{key}{A string. Serves as a unique key for the database source which
will identify data in the destination DB as coming from a specific source.}

\item{desc}{A string. The description the will be used to describe the source
database contents which are imported into the Postgres destination.}
}
\value{
A list of schema names, and within each a list of tibbles containing
source data.
}
\description{
Retrieves data tables from data source and returns them as a list of tibbles.
}
