% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R, R/lpi_calc.R, R/lpi_convert.R
\name{\%do\%}
\alias{\%do\%}
\title{load special binary operators}
\usage{
obj \%do\% ex

obj \%do\% ex

obj \%do\% ex
}
\description{
load special binary operators

load special binary operators

load special binary operators
}
