% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_calc.R
\name{import.data}
\alias{import.data}
\title{This function connects to a Postgres instance of the LLEIA database and reads
in relevant tables for Point intercept calculations.}
\usage{
import.data(dbname, host, port, user, password)
}
\arguments{
\item{dbname}{A string. The database name for the postgres connection.}

\item{host}{A string. The IP address or DNS name for the postgres connection}

\item{port}{An integer. The port for the postres connection.}

\item{user}{A string. The user name for the postgres connection}

\item{password}{A string. The password for the postgres connection.}
}
\value{
A list of tibbles imported from the connection source.
}
\description{
This function connects to a Postgres instance of the LLEIA database and reads
in relevant tables for Point intercept calculations.
}
