% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.to.post}
\alias{import.to.post}
\title{The main processing function used to import data from source data into a
PostGIS database}
\usage{
import.to.post(
  dbname,
  user,
  password,
  src.path,
  key,
  desc,
  host = "localhost",
  port = 5432,
  update = FALSE,
  log = FALSE,
  save.raw = NULL,
  save.processed = NULL,
  verbose = FALSE,
  chunk.size = 1000,
  skip.refresh = FALSE
)
}
\arguments{
\item{dbname}{A string. The database name to connect to in the postgres
instance.}

\item{user}{A string. The database user used to connect to the database.}

\item{password}{A string. The password used to connect to the database.}

\item{src.path}{A string file path pointing to the source database or saved
RDS file. File name extensions must be one of (.mdb, .accdb, .sqlite, .db,
.rds) and folder paths must have the extension of (.gdb).}

\item{key}{A string. Serves as a unique key for the database source which
will identify data in the destination DB as coming from a specific source.}

\item{desc}{A string. The description the will be used to describe the source
database contents which are imported into the Postgres destination.}

\item{host}{A string. The IP address or DNS name which hosts the database.}

\item{port}{An integer. The port which the postgres service monitors for
connections.}

\item{update}{logical. A flag that instructs the function to construct an
UPSERT statement instead of an INSERT statement.}

\item{log}{logical. A flag that tells the function to create a log file of
the import, saved in the current working directory in the format of
import_YYYMMDDHHMMSS.log}

\item{save.raw}{A string file path to which to save data imported data from
\code{src.path} via \code{get.src.tables()}. This file can be used in
as a future \code{src.path}. Primarily used for testing purposes.}

\item{save.processed}{A string file path to which to save data imported data
from \code{src.path} via \code{get.src.tables()} which has been processed
via \code{process.terradat()}. This file can be used in
as a future \code{src.path}. Primarily used for testing purposes.}

\item{verbose}{logical. A flag telling the function to be more verbose in
its messaging.}

\item{chunk.size}{An integer.This tells the function how many rows to
attempt to insert at once. Failure on a chunk will cause the function
to default to row-wise inserts for the entire chunk.}

\item{skip.refresh}{logical. A flag telling the function to skip refreshing
materialized views. This can be helpful in speeding up the importing of
multiple databases.}
}
\description{
The main processing function used to import data from source data into a
PostGIS database
}
