% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{insert.order}
\alias{insert.order}
\title{Recursive function which returns an ordered table of schema table names
where tables that are foreign key parents are listed in order before their
children}
\usage{
insert.order(schema, level = 0, processed.tables = NULL, tables = NULL)
}
\arguments{
\item{schema}{A string. The schema to be queried.}

\item{level}{An integer. An internal parameter that the function uses to keep
track of how many times it has self called.}

\item{processed.tables}{A tibble. An internal parameter that the function
uses to keep track of which tables have been processed and stored in the
return variable.}

\item{tables}{A string vector. An internal parameter that the function uses
to keep keep track of unprocessed table names.}
}
\value{
A tibble that gives each table in \code{schema} and a proper insert
order that will prevent improper foreign key violations.
}
\description{
Recursive function which returns an ordered table of schema table names
where tables that are foreign key parents are listed in order before their
children
}
