% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{insert.row}
\alias{insert.row}
\title{Used to insert individual rows via apply() family function or iterative loop}
\usage{
insert.row(row, stmt, log = NULL)
}
\arguments{
\item{row}{A named list or tibble containing the values to bind to the INSERT
statement via parameter substitution.}

\item{stmt}{A string containing the INSERT statement with named parameters to
use with glue::glue syntax (e.g. {my.parameter}).}

\item{log}{A string file path location of the log file to write
results to.}
}
\value{
The number of rows affected by the insert.
}
\description{
Used to insert individual rows via apply() family function or iterative loop
}
