% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_calc.R
\name{load.indicators}
\alias{load.indicators}
\title{This function reads in a tab delimited file containing filter definitions for
multiple indicators and returns a tibble version of it.}
\usage{
load.indicators(indicators)
}
\arguments{
\item{indicators}{A string file path which points to a tab delimited file
(with field names) containing three fields: 1) filter.tbl - this is either
"long" or "wide" and indicates which type of point intercept table is used
in calculation of the indicator. 2) name - this is the name of the
indicator. It is recommended for later processing that these name contain
only alphanumeric characters or underscores. 3) filter.exp - a
\code{dplyr::filter()} expression with the first argument = "hits" (e.g.
dplyr::filter(hits, my_field = "some parameter"). Multiple filter
expressions can by union-ed with the \code{union} operator, and the filter
statement can use the (now-deprecated) \code{filter_if()},
\code{filter_at}, and \code{filter_all} as well as the use the
\code{across()} operator.}
}
\value{
A tibble containing the filter statements from \code{indicators}.
}
\description{
This function reads in a tab delimited file containing filter definitions for
multiple indicators and returns a tibble version of it.
}
