% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_convert.R
\name{lpi.to.gis}
\alias{lpi.to.gis}
\title{The main processing function for the module.}
\usage{
lpi.to.gis(
  dbname,
  user,
  password,
  in.path,
  out.path,
  host = "localhost",
  port = 5432,
  layer.name = "point",
  srid = 4326,
  sep = ",",
  cover.str = NULL,
  height.str = NULL,
  dead.str = NULL
)
}
\arguments{
\item{dbname}{A string. The database name to connect to in the postgres
instance.}

\item{user}{A string. The database user used to connect to the database.}

\item{password}{A string. The password used to connect to the database.}

\item{in.path}{A string. The file path to am RDS or CSV file produced from
the lpi_calc module.}

\item{out.path}{A string which denotes the file path to write the output to.
The extension of out.path is used by sf::st_write() to guess the format
of the output using available drivers in sf::st_drivers().}

\item{host}{A string. The IP address or DNS name which hosts the database.}

\item{port}{An integer. The port which the postgres service monitors for
connections.}

\item{layer.name}{A string. The name of the layer to be used in the
destination (\code{out.path}) when the feature is written.}

\item{srid}{An integer. The SRID/EPSG code that the output should have.}

\item{sep}{A character which is used to delimit the input file
(\code{in.path}) in the case that it is in the CSV format.}

\item{cover.str}{A string in the format of "{indicator1}, {hit_type1};
{indicator2}, {hit_type2}; etc." which is used to build filters for in.tbl
and determines which cover fields the wide output table will have.}

\item{height.str}{A string. Similar to \code{cover.str}, but for height fields.}

\item{dead.str}{A string. Similar to \code{cover.str}, but for dead fields.}
}
\value{
Nothing.
}
\description{
The main processing function for the module.
}
