% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{msg.out}
\alias{msg.out}
\title{This function will take a log message and write it to stdout and/or a log
file.}
\usage{
msg.out(msg, log = NULL, sep = "\\n", prnt = TRUE)
}
\arguments{
\item{msg}{A string. The message to write.}

\item{log}{A string file path to a log wehre messages are to be written.}

\item{sep}{A string, the separator to paste after the message (e.g. an EOL
character)}

\item{prnt}{logical. A flag indicating whether or not to print the msg to
stdout}
}
\description{
This function will take a log message and write it to stdout and/or a log
file.
}
