% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{post.to.sqlite}
\alias{post.to.sqlite}
\title{The main processing function used to copy PostGIS tables to the SQLite
database.}
\usage{
post.to.sqlite(
  pg_db,
  pg_user,
  password,
  dbpath,
  port = 5432,
  host = "localhost",
  schema = "eco"
)
}
\arguments{
\item{pg_db}{A string. The database to connect to in the Postgres instance.}

\item{pg_user}{A string. The user name to connect to the Postgres instance
with.}

\item{password}{A string. The password to connect to the Postgres instance
with.}

\item{dbpath}{A string file path that points to the location to create and
store data from the Postgres \code{schema} in an SQLite/SpatiaLite
database.}

\item{port}{An integer. The port which the Postgres instance uses to monitor
connection requests.}

\item{host}{A string. The IP address or DNS name which hosts the database.}

\item{schema}{A string. The name of the schema in the Postgres instance.}
}
\description{
The main processing function used to copy PostGIS tables to the SQLite
database.
}
