% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{process.terradat}
\alias{process.terradat}
\title{Data in TerrAdat has some specials cases, namely that many of the fields,
including plotkey, linekey, etc are too long for DIMA due to some of the
source data having been converted from another schema type to resemble DIMA
data. We are forced to rekey or truncate those fields, as one of the goals
of constructing LLEIA in the manner it is, is to be able to "re-DIMA" the
enterprise level data into a blank DIMA if we so chose. This function looks
at the source data and performs some checks and conversions if it is found to
be from a TerrAdat source.}
\usage{
process.terradat(imported)
}
\arguments{
\item{imported}{A list of lists of tibbles, constructed via
\code{get.src.tables()}}
}
\value{
A processed list of lists of tibbles, which is compatible with the
DIMA schema, as well as other data such as database keys found in TerrAdat,
and other values (source path, md5hash, key, description and table name
list)
}
\description{
Data in TerrAdat has some specials cases, namely that many of the fields,
including plotkey, linekey, etc are too long for DIMA due to some of the
source data having been converted from another schema type to resemble DIMA
data. We are forced to rekey or truncate those fields, as one of the goals
of constructing LLEIA in the manner it is, is to be able to "re-DIMA" the
enterprise level data into a blank DIMA if we so chose. This function looks
at the source data and performs some checks and conversions if it is found to
be from a TerrAdat source.
}
