% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{refresh.views}
\alias{refresh.views}
\title{A recursive function which can be used to determine the proper order in which
to refresh materialized views, i.e. materialized views which depend on other
materialized views will be refreshed after the view it depends on is
refreshed.}
\usage{
refresh.views(done = character(0), level = 0)
}
\arguments{
\item{done}{A string vector. An internal parameter which the function uses to
keep track of which views have already been refreshed.}

\item{level}{An integer. An internal parameter which the function uses to
keep track of how many times it has been called.}
}
\description{
A recursive function which can be used to determine the proper order in which
to refresh materialized views, i.e. materialized views which depend on other
materialized views will be refreshed after the view it depends on is
refreshed.
}
