% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{rowwise.insert}
\alias{rowwise.insert}
\title{Inserts the source data in a source table into the Postgres destination
using row-wise inserts (one row at a time.)  This is slower than a multi-row
parameter bind, but allows the insert process to capture single row insert
errors, log them, and continue processing.}
\usage{
rowwise.insert(update, schema, table, table.name, cols, cast, log)
}
\arguments{
\item{update}{logical. A flag that instructs the function to construct an
UPSERT statement instead of an INSERT statement.}

\item{schema}{A string. The name of the schema in the destination.}

\item{table}{A tibble containing the source data to be inserted.}

\item{table.name}{A name of the table in the destination for which to build
the INSERT statement.}

\item{cols}{A string vector containing the names of columns to use for
the insert statement.}

\item{cast}{A logical vector telling the function whether to encapsulate the
parameter in the SQL in a CAST function, hopefully mitigating type errors.}

\item{log}{A string file path location of the log file to write
results to.}
}
\value{
integer. Total number of rows affected by the insert.
}
\description{
Inserts the source data in a source table into the Postgres destination
using row-wise inserts (one row at a time.)  This is slower than a multi-row
parameter bind, but allows the insert process to capture single row insert
errors, log them, and continue processing.
}
