% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sql.list}
\alias{sql.list}
\title{Database creation SQL converted from raw SQL into list format for use in
create.lleiadb function.}
\format{
A list with one element for each SQL file. Each sub-list has two
character vectors:
\describe{
\item{sql}{A vector of SQL statements, suitable for executing within the
database}
\item{formatted}{A vector of formatted SQL statements suitable for
printing but not executing.}
}
}
\usage{
sql.list
}
\description{
Database creation SQL converted from raw SQL into list format for use in
create.lleiadb function.
}
\keyword{datasets}
