% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_calc.R
\name{test.indicators}
\alias{test.indicators}
\title{This function will filter out a table based on a string evaluated as an
expression.}
\usage{
test.indicators(name, filter.exp, calc.type, hits)
}
\arguments{
\item{name}{A string. The name of the indicator.}

\item{filter.exp}{A string. A \code{dpylr::filter()} expression (or union-ed
set thereof) which is used to filter the table}

\item{calc.type}{A string. The table type on which to perform the filter
expression.}

\item{hits}{A tibble. The table to perform the filter on.}
}
\value{
A tibble, with the \code{filter.exp} applied to \code{hits}
}
\description{
This function will filter out a table based on a string evaluated as an
expression.
}
\examples{
test.indicators("tree", 'dplyr::filter(hits, growth_habit_first ==
  "Tree", "long", pintercept.long')
}
