% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{try.execute.sql}
\alias{try.execute.sql}
\title{Attempts to execute SQL statement on a db connection and prints/logs the
SQL statement and error to stdout and/or log file in case of failure.}
\usage{
try.execute.sql(dbcon, sql)
}
\arguments{
\item{dbcon}{A database connection created with DBI::dbConnect().}

\item{sql}{character string. An sql statement to execute.}
}
\value{
A scalar numeric that specifies the number of rows affected by the
statement.
}
\description{
Attempts to execute SQL statement on a db connection and prints/logs the
SQL statement and error to stdout and/or log file in case of failure.
}
