% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{vdigest}
\alias{vdigest}
\title{This vectorizes the digest function from the digest package}
\usage{
vdigest(
  object,
  algo = c("md5", "sha1", "crc32", "sha256", "sha512", "xxhash32", "xxhash64",
    "murmur32", "spookyhash", "blake3"),
  serialize = TRUE,
  file = FALSE,
  length = Inf,
  skip = "auto",
  ascii = FALSE,
  raw = FALSE,
  seed = 0,
  errormode = c("stop", "warn", "silent"),
  serializeVersion = .getSerializeVersion()
)
}
\description{
This vectorizes the digest function from the digest package
}
