% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_convert.R
\name{write.spatial}
\alias{write.spatial}
\title{Takes a tibble produced from filter.table(), joins site data and point data
and writes to a spatial data set based on the output file extension.}
\usage{
write.spatial(
  tbl.out,
  out.path,
  dbname,
  user,
  password,
  host,
  port,
  layer.name,
  srid
)
}
\arguments{
\item{tbl.out}{A tibble produced by the filter.table() function.}

\item{out.path}{A string which denotes the file path to write the output to.
The extension of out.path is used by sf::st_write() to guess the format
of the output using available drivers in sf::st_drivers().}

\item{dbname}{A string. The database name to connect to in the postgres
instance.}

\item{user}{A string. The database user used to connect to the database.}

\item{password}{A string. The password used to connect to the database.}

\item{host}{A string. The IP address or DNS name which hosts the database.}

\item{port}{An integer. The port which the postgres service monitors for
connections.}

\item{layer.name}{A string. The name of the layer to be used in the
destination (\code{out.path}) when the feature is written.}

\item{srid}{An integer. The SRID/EPSG code that the output should have.}
}
\description{
Takes a tibble produced from filter.table(), joins site data and point data
and writes to a spatial data set based on the output file extension.
}
