% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create.lleiadb}
\alias{create.lleiadb}
\title{The main processing function which creates the PostGIS database and populates
it with default data.}
\usage{
create.lleiadb(
  dbname,
  user,
  password,
  sql.path,
  spatial.path,
  host = "localhost",
  port = 5432
)
}
\arguments{
\item{dbname}{A string. The database name to connect to in the postgres
instance.}

\item{user}{A string. The database user used to connect to the database.}

\item{password}{A string. The password used to connect to the database.}

\item{sql.path}{A string directory path which points to the location where
the SQL files for DB creation are stored.}

\item{spatial.path}{A string directory path which points to the location
where spatial data to be imported into the PostGIS instance is stored.}

\item{host}{A string. The IP address or DNS name which hosts the database.}

\item{port}{An integer. The port which the postgres service monitors for
connections.}
}
\value{
Nothing.
}
\description{
The main processing function which creates the PostGIS database and populates
it with default data.
}
