% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_db.R
\name{create.pool}
\alias{create.pool}
\title{Creates a pooled database object that other functions use to interact with
the database.}
\usage{
create.pool(dbname, host = "localhost", port = 5432, user, password)
}
\arguments{
\item{dbname}{A character vector. The database name to connect to in the postgres
instance.}

\item{host}{A character vector. The IP address or DNS name which hosts the database.}

\item{port}{An integer. The port which the postgres service monitors for
connections.}

\item{user}{A character vector. The database user used to connect to the database.}

\item{password}{A character vector. The password used to connect to the database.}
}
\description{
Creates a pooled database object that other functions use to interact with
the database.
}
