% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi_convert.R
\name{lpi.to.gis}
\alias{lpi.to.gis}
\title{Takes an output from \code{calc.lpi}, coverts it to a gis friendly wide
format, and adds geomtery fields.}
\usage{
lpi.to.gis(
  dbname,
  user,
  password,
  in.path,
  out.path,
  host = "localhost",
  port = 5432,
  layer.name = "point",
  srid = 4326,
  sep = ",",
  sd = TRUE,
  n = TRUE,
  filter.df = tibble::tibble(field = character(), indicator = character(), hit_type =
    character())
)
}
\arguments{
\item{dbname}{A string. The database name to connect to in the postgres
instance.}

\item{user}{A string. The database user used to connect to the database.}

\item{password}{A string. The password used to connect to the database.}

\item{in.path}{A string. The file path to am RDS or CSV file produced from
the lpi_calc module.}

\item{out.path}{A string which denotes the file path to write the output to.
The extension of out.path is used by sf::st_write() to guess the format
of the output using available drivers in sf::st_drivers().}

\item{host}{A string. The IP address or DNS name which hosts the database.}

\item{port}{An integer. The port which the postgres service monitors for
connections.}

\item{layer.name}{A string. The name of the layer to be used in the
destination (\code{out.path}) when the feature is written.}

\item{srid}{An integer. The SRID/EPSG code that the output should have.}

\item{sep}{A character which is used to delimit the input file
(\code{in.path}) in the case that it is in the CSV format.}

\item{sd}{Boolean. Use standard deviation field(s) in wide result.}

\item{n}{Boolean. Use line count (n) field(s) in wide result.}

\item{filter.df}{A tibble with the character fields: \code{field},
\code{indicator}, and \code{hit_type}. The \code{field} column must contain
one of \code{c("cover", "height", "dead")}. The indicator field must match
indicators to filter \code{in.tbl} by or NA, which will result in all
indicators of a certain \code{hit_type} being selected. The same logic
applies to the \code{hit_type} field.}
}
\value{
Nothing.
}
\description{
Takes an output from \code{calc.lpi}, coverts it to a gis friendly wide
format, and adds geomtery fields.
}
